+++ include/mruby_integration/core/input.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_core_input(mrb_state*);