+++ mrb_doc/core/input/gamepad.rb
@@ -1,9 +1,0 @@
-# Setup gamepad mappings using the
-# [SDL_GameControllerDB](https://github.com/gabomdq/SDL_GameControllerDB) format.
-# @param mappings [String]
-# @return [Integer]
-def set_gamepad_mappings(mappings)
-  # mrb_set_gamepad_mappings
-  # src/mruby_integration/core/input/gamepad.cpp
-  1
-end