+++ src/mruby_integration/core.cpp
@@ -3,7 +3,6 @@
 #include "mruby_integration/core/drawing.hpp"
 #include "mruby_integration/core/files.hpp"
 #include "mruby_integration/core/gestures.hpp"
-#include "mruby_integration/core/input.hpp"
 #include "mruby_integration/core/misc.hpp"
 #include "mruby_integration/core/timing.hpp"
 
@@ -15,7 +14,6 @@ append_core(mrb_state* mrb)
   append_core_drawing(mrb);
   append_core_files(mrb);
   append_core_gestures(mrb);
-  append_core_input(mrb);
   append_core_misc(mrb);
   append_core_timing(mrb);